requirements.txt
beautifulsoup4==4.12.3
scrapy==2.11.0
requests==2.31.0
wget==3.2
python-docx==1.1.0
html2text==2024.2.26
markdown==3.5.2
Pillow==10.2.0
pyyaml==6.0.1
lxml==5.1.0